Add PurgeRecording and many bug fixes.
It turns out that the relevantId returned by AddRecording is not the
recordingId. This cascades into significant changes in the tests and
examples.

TestListRecordingsFromUri is now consistently working ans does a
better job of checking as it goes.

Added PurgeRecording so we can clean up after ourselves in the tests.

Initial attempt at making Debug mode a little liess verbose and logging
at Info is now more sparse.

Fixed a bug where RecordingSignals could interrupt the flow of control
messages as we weren't asking for enough fragments.

Fixed timing bugs on various protocol requests by using the archive
poll() which ensures operations complete.

Fixed not clearing old state in the control Results so that single
tests/cases would work but were not repeatable.