Naming consistency part #2
For a protocol message XxxRequest, the proxy has a function XxxRequest() which calls an encoder function XxxPacket().

The archive mostly has an Xxx() but not uniformly as we sometimes
choose to have the function named as for Java where that name more
clearly matches the semantics of the protocol message.