API additions, PollForResponse() returns a value, add some documentation
Add to archive the API additions:
 * BoundedReplay
 * TruncateRecording
 * GetStartPosition
 * GetStopPosition
 * FindLastMatchingRecording
 * ListRecordingSubscriptions() with tests
 * DetachSegments
 * DeleteDetachedSegments
 * PurgeSegments
 * AttachSegments
 * MigrateSegments

Restructure around having PollForResponse() return an int64. This is
used (and sometimes abused as a boolean) in a variety of places to
simplify the archive's use of control.

Revise and improve some API documentation, and better document some
design decisions around sync/async, locking, and naming.