Tweak the fragmentCount for ControlResponses.
Set this to ten (matching Java) so we handle any RecordingSignals that
appear at the same time.

It's not definite this will always work as there may be timing issues
but in the general instance this will work as expected.

Add a test case which also uses and tests the RecordingEvents which is
a reliable way of checking for changes to recordings.