Refactor/modernize some tests
Pull out logbuffers_test's test file/logbuffer constructor so it is a
separate helper function.  Change publication_test to use it also.

Modernize publication_test and logbuffers_test so we get cleaner code
and cleaner error messages.

Add an interface to subscription.go so it can have an injected mock/fake
ClientConductor.  Start to lay the groundwork for a fake Image
dependency injection, but it is very rudimentary.

Unfortunately Image is used too widely, and with such
performance-specific tweaks, that it is not conducive to being
refactored to make it easier to inject a mock.  Instead we'll have to
deal with fakes.  The logbuffer refactorings above help with this.