Fix bug in sys_test
It called Poll() with a fragment limit of 10, but only incremented the
received counter if Poll() returned exactly one fragment.  If >=2
arrived, it would spin and timeout, incorrectly failing.

I changed the tests to use suite require/expect calls to make the errors
cleaner.

I also lowered the send timeout from 50ms to 1ms.  That was forcing this
test to take 5s, and with this bug fixed, I don't think there's any more
need to wait between sends.