Subscriber error handling
Changed FragmentHandler so it returns an error.  That error gets
propagated up through Image.poll(), and will interrupt a batch poll.
This is the functionality in Java Aeron.

Added an optional subscriber-specific error handler, mirroring the Java
impl.

Note that this is fully functional, but I wouldn't call it done.  There
are a few existing places where a logged error should be converted to a
returned error.  I should also make the equivalent change in the
ControlledFragmentHandler.  However, I wanted to get initial feedback on
this approach, especially considering that it breaks a public API
(although in a tiny way - adding "return nil" fixes the break)