Address review comments
Add temporary error handling to the Deprecated calls.

Add a panic

claimCapacity is back to returning negative sentinel values.  Also made
the analagous revert elsewhere to keep with the theme of keeping
Java/C++ sentinel returns, vs. named error messages.

In the theme of not minding panics if 1) returning an error is an
annoying API change and 2) the panic represents a serious configuration
problem, I reverted one or two other changes.